+++ mojo/__init__.py
@@ -1,3 +1,3 @@
-__version__ = "1.2.22"
+__version__ = "1.2.23"
 
 from mojo.helpers.response import JsonResponse